tests/VnV/probes/mqtt/requirements.txt
paho-mqtt==1.3.0